requirements.txt
numpy
dicttoxml
pynrrd
opencv-python-headless
